+++ saleor/graphql/webhook/subscription_types.py
@@ -23,7 +23,7 @@ from ...webhook.event_types import WebhookEventAsyncType
 from ..account.types import User as UserType
 from ..app.types import App as AppType
 from ..channel import ChannelContext
-from ..core.descriptions import ADDED_IN_32, ADDED_IN_34, PREVIEW_FEATURE
+from ..core.descriptions import ADDED_IN_32, ADDED_IN_34, ADDED_IN_35, PREVIEW_FEATURE
 from ..core.scalars import PositiveDecimal
 from ..payment.enums import TransactionActionEnum
 from ..payment.types import TransactionItem
@@ -70,6 +70,9 @@ class Event(graphene.Interface):
     @classmethod
     def get_type(cls, object_type: str):
         types = {
+            WebhookEventAsyncType.ADDRESS_CREATED: AddressCreated,
+            WebhookEventAsyncType.ADDRESS_UPDATED: AddressUpdated,
+            WebhookEventAsyncType.ADDRESS_DELETED: AddressDeleted,
             WebhookEventAsyncType.APP_INSTALLED: AppInstalled,
             WebhookEventAsyncType.APP_UPDATED: AppUpdated,
             WebhookEventAsyncType.APP_DELETED: AppDeleted,
@@ -130,6 +133,9 @@ class Event(graphene.Interface):
             WebhookEventAsyncType.PAGE_CREATED: PageCreated,
             WebhookEventAsyncType.PAGE_UPDATED: PageUpdated,
             WebhookEventAsyncType.PAGE_DELETED: PageDeleted,
+            WebhookEventAsyncType.PAGE_TYPE_CREATED: PageTypeCreated,
+            WebhookEventAsyncType.PAGE_TYPE_UPDATED: PageTypeUpdated,
+            WebhookEventAsyncType.PAGE_TYPE_DELETED: PageTypeDeleted,
             WebhookEventAsyncType.SHIPPING_PRICE_CREATED: ShippingPriceCreated,
             WebhookEventAsyncType.SHIPPING_PRICE_UPDATED: ShippingPriceUpdated,
             WebhookEventAsyncType.SHIPPING_PRICE_DELETED: ShippingPriceDeleted,
@@ -170,6 +176,33 @@ class Event(graphene.Interface):
         if isinstance(info.context.requestor, AnonymousUser):
             return None
         return info.context.requestor
+
+
+class AddressBase(AbstractType):
+    address = graphene.Field(
+        "saleor.graphql.account.types.Address",
+        description="The address the event relates to." + ADDED_IN_35 + PREVIEW_FEATURE,
+    )
+
+    @staticmethod
+    def resolve_address(root, _info):
+        _, address = root
+        return address
+
+
+class AddressCreated(ObjectType, AddressBase):
+    class Meta:
+        interfaces = (Event,)
+
+
+class AddressUpdated(ObjectType, AddressBase):
+    class Meta:
+        interfaces = (Event,)
+
+
+class AddressDeleted(ObjectType, AddressBase):
+    class Meta:
+        interfaces = (Event,)
 
 
 class AppBase(AbstractType):
@@ -730,6 +763,35 @@ class PageDeleted(ObjectType, PageBase):
         interfaces = (Event,)
 
 
+class PageTypeBase(AbstractType):
+    page_type = graphene.Field(
+        "saleor.graphql.page.types.PageType",
+        description="The page type the event relates to."
+        + ADDED_IN_35
+        + PREVIEW_FEATURE,
+    )
+
+    @staticmethod
+    def resolve_page_type(root, _info):
+        _, page_type = root
+        return page_type
+
+
+class PageTypeCreated(ObjectType, PageTypeBase):
+    class Meta:
+        interfaces = (Event,)
+
+
+class PageTypeUpdated(ObjectType, PageTypeBase):
+    class Meta:
+        interfaces = (Event,)
+
+
+class PageTypeDeleted(ObjectType, PageTypeBase):
+    class Meta:
+        interfaces = (Event,)
+
+
 class ShippingPriceBase(AbstractType):
     shipping_method = graphene.Field(
         "saleor.graphql.shipping.types.ShippingMethodType",
@@ -963,6 +1025,9 @@ class Subscription(ObjectType):
 # so all subscription events types need to be added manually to `Schema`.
 
 SUBSCRIPTION_EVENTS_TYPES = [
+    AddressCreated,
+    AddressUpdated,
+    AddressDeleted,
     AppInstalled,
     AppUpdated,
     AppDeleted,
@@ -1019,6 +1084,9 @@ SUBSCRIPTION_EVENTS_TYPES = [
     PageCreated,
     PageUpdated,
     PageDeleted,
+    PageTypeCreated,
+    PageTypeUpdated,
+    PageTypeDeleted,
     ShippingPriceCreated,
     ShippingPriceUpdated,
     ShippingPriceDeleted,