+++ saleor/plugins/webhook/tests/subscription_webhooks/test_create_deliveries_for_subscription.py
@@ -16,6 +16,7 @@ from .....webhook.event_types import WebhookEventAsyncType
 from ...tasks import create_deliveries_for_subscriptions, logger
 from . import subscription_queries
 from .payloads import (
+    generate_address_payload,
     generate_app_payload,
     generate_category_payload,
     generate_collection_payload,
@@ -26,6 +27,7 @@ from .payloads import (
     generate_menu_item_payload,
     generate_menu_payload,
     generate_page_payload,
+    generate_page_type_payload,
     generate_sale_payload,
     generate_shipping_method_payload,
     generate_voucher_created_payload_with_meta,
@@ -70,6 +72,57 @@ def test_subscription_query_with_meta(
     assert deliveries[0].webhook == webhooks[0]
 
 
+def test_address_created(address, subscription_address_created_webhook):
+    # given
+    webhooks = [subscription_address_created_webhook]
+    event_type = WebhookEventAsyncType.ADDRESS_CREATED
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, address, webhooks)
+
+    # then
+    expected_payload = json.dumps({"address": generate_address_payload(address)})
+
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
+def test_address_updated(address, subscription_address_updated_webhook):
+    # given
+    webhooks = [subscription_address_updated_webhook]
+    event_type = WebhookEventAsyncType.ADDRESS_UPDATED
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, address, webhooks)
+
+    # then
+    expected_payload = json.dumps({"address": generate_address_payload(address)})
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
+def test_address_deleted(address, subscription_address_deleted_webhook):
+    # given
+    webhooks = [subscription_address_deleted_webhook]
+
+    id = address.id
+    address.delete()
+    address.id = id
+
+    event_type = WebhookEventAsyncType.ADDRESS_DELETED
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, address, webhooks)
+
+    # then
+    expected_payload = json.dumps({"address": generate_address_payload(address)})
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
 def test_app_installed(app, subscription_app_installed_webhook):
     # given
     webhooks = [subscription_app_installed_webhook]
@@ -90,13 +143,13 @@ def test_app_updated(app, subscription_app_updated_webhook):
     # given
     webhooks = [subscription_app_updated_webhook]
     event_type = WebhookEventAsyncType.APP_UPDATED
-    gift_card_id = graphene.Node.to_global_id("App", app.id)
+    app_id = graphene.Node.to_global_id("App", app.id)
 
     # when
     deliveries = create_deliveries_for_subscriptions(event_type, app, webhooks)
 
     # then
-    expected_payload = generate_app_payload(app, gift_card_id)
+    expected_payload = generate_app_payload(app, app_id)
     assert deliveries[0].payload.payload == expected_payload
     assert len(deliveries) == len(webhooks)
     assert deliveries[0].webhook == webhooks[0]
@@ -1120,6 +1173,55 @@ def test_page_deleted(page, subscription_page_deleted_webhook):
 
     # when
     deliveries = create_deliveries_for_subscriptions(event_type, page, webhooks)
+
+    # then
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
+def test_page_type_created(page_type, subscription_page_type_created_webhook):
+    # given
+    webhooks = [subscription_page_type_created_webhook]
+    event_type = WebhookEventAsyncType.PAGE_TYPE_CREATED
+    expected_payload = json.dumps(generate_page_type_payload(page_type))
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, page_type, webhooks)
+
+    # then
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
+def test_page_type_updated(page_type, subscription_page_type_updated_webhook):
+    # given
+    webhooks = [subscription_page_type_updated_webhook]
+    event_type = WebhookEventAsyncType.PAGE_TYPE_UPDATED
+    expected_payload = json.dumps(generate_page_type_payload(page_type))
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, page_type, webhooks)
+
+    # then
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
+def test_page_type_deleted(page_type, subscription_page_type_deleted_webhook):
+    # given
+    page_type_id = page_type.id
+    page_type.delete()
+    page_type.id = page_type_id
+
+    webhooks = [subscription_page_type_deleted_webhook]
+    event_type = WebhookEventAsyncType.PAGE_TYPE_DELETED
+    expected_payload = json.dumps(generate_page_type_payload(page_type))
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(event_type, page_type, webhooks)
 
     # then
     assert deliveries[0].payload.payload == expected_payload